Changed contract of search namespace to match reduce.
This is a backward-incompatible change in the `search` namespace. Previously,
the functions took arguments either in order `[input next-vals-fn pred]` for stateless calls or
`[initial-state initial-vals eval-fn]`. This did not match the order of arguments in the
`reduce` function. So even though it's unconventional, this is a contract-changing modification
to reorder the arguments. Now the stateless function arguments go `[pred next-vals-fn input]`
and the stateful ones use `[eval-fn initial-state initial-vals]`.